pyproject.toml
[project]
name = "nami-evals"
version = "0.1.0"
description = "Geração de conversas sintéticas multi-turno para avaliação de prompts NAMI"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.50.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "python-docx>=1.2.0",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "sse-starlette>=1.6",
]

[project.scripts]
nami-evals = "src.runner:app"
nami-playground = "src.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
